requirements.txt
tensorflow==2.3.0
tensorflow-hub==0.9.0
numpy==1.18.5
Pillow==7.2.0
requests==2.24.0
google-cloud-storage==1.16.1
tweepy==3.9.0
